Move string literal decoding into `semantic::types`
This allows precise typing of literal string constants.
