+++ crates/solidity-v2/outputs/cargo/semantic/src/types/literals/mod.rs
@@ -1,1 +1,6 @@
 pub(crate) mod numbers;
+pub(crate) mod strings;
+
+pub use strings::{
+    value_of_hex_string_literals, value_of_string_literals, value_of_unicode_string_literals,
+};
