+++ crates/solidity-v2/outputs/cargo/semantic/src/types/registry.rs
@@ -229,7 +229,7 @@ impl TypeRegistry {
             (
                 Type::Literal(LiteralKind::HexString { bytes } | LiteralKind::String { bytes }),
                 Type::ByteArray { width },
-            ) if *bytes == *width => true,
+            ) if *bytes == *width as usize => true,
 
             (
                 Type::Array {
